fargate2/requirements.txt
appdirs==1.4.4
boto3==1.17.23
botocore==1.20.23
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
filelock==3.0.12
idna==2.10
jmespath==0.10.0
mysql-connector==2.2.9
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.25.10
virtualenv==20.2.2